+++ src/assorthead/__init__.py
@@ -15,4 +15,4 @@ except PackageNotFoundError:  # pragma: no cover
 finally:
     del version, PackageNotFoundError
 
-from .includes import includes
+from .includes import includes, licenses