+++ src/assorthead/includes.py
@@ -4,10 +4,17 @@ from typing import List
 
 
 def includes() -> str:
-    """Provides access to assorted C++ headers for downstream packages.
-
-    Returns:
-        str: Path to a directory containing lots of header files.
+    """
+    Returns: Path to a directory containing lots of header files.
     """
     dirname = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
     return os.path.join(dirname, "include")
+
+
+def licenses() -> str:
+    """
+    Returns: Path to a directory containing licenses for each library.
+    """
+    dirname = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
+    return os.path.join(dirname, "licenses")
+